Refactor functions for paper; #1033
